version incompatibility

+++ hw_asr/utils/util.py
@@ -58,7 +58,7 @@ def prepare_device(n_gpu_use):
 class MetricTracker:
     def __init__(self, *keys, writer=None):
         self.writer = writer
-        self._data = pd.DataFrame(index=keys, columns=["total", "counts", "average"])
+        self._data = pd.DataFrame(index=keys, columns=["total", "counts", "average"], dtype = float)
         self.reset()
 
     def reset(self):
